iceberg: lift to_uri to metadata_io
And avoid hardcoded default scheme in make_uri.

(cherry picked from commit 13d0cdc94bf55293dcd75a13bf2f9c0e31f21a86)
