+++ src/v/iceberg/manifest_io.cc
@@ -94,10 +94,6 @@ manifest_io::upload_manifest_list(const uri& uri, const manifest_list& m) {
     return upload_manifest_list(manifest_list_path(path_res.value()), m);
 }
 
-uri manifest_io::to_uri(const std::filesystem::path& p) const {
-    return make_uri(bucket_, p);
-}
-
 checked<std::filesystem::path, metadata_io::errc>
 manifest_io::from_uri(const uri& s) const {
     try {
